code/requirements.txt
wandb
matplotlib
seaborn
nibabel
torchio
torchsummary
Ipython
pandas
numpy
torchvision
monai
monai-weekly
omegaconf
timm
scikit-learn
torchprofile
medio
medpy
opencv-python
scikit-image
synapseclient
blobfile